文章

+++ living/living/Model/Request/LMFriendListRequest.m
@@ -10,16 +10,16 @@
 
 @implementation LMFriendListRequest
 
--(id)initWithPageIndex:(int)pageIndex andPageSize:(int)pageSize
+-(id)initWithPageIndex:(NSInteger)pageIndex andPageSize:(NSInteger)pageSize
 {
     self = [super init];
     if (self) {
         NSMutableDictionary *body = [NSMutableDictionary new];
         if (pageIndex != -1) {
-            [body setObject:[NSString stringWithFormat:@"%d", pageIndex] forKey:@"pageIndex"];
+            [body setObject:[NSString stringWithFormat:@"%ld", (long)pageIndex] forKey:@"pageIndex"];
         }
         if (pageSize != -1) {
-            [body setObject:[NSString stringWithFormat:@"%d", pageSize] forKey:@"pageSize"];
+            [body setObject:[NSString stringWithFormat:@"%ld", (long)pageSize] forKey:@"pageSize"];
         }
         
         NSMutableDictionary *parmDic = [self params];
